==> Pandas in python/1. Pandas.ipynb <==
# Pandas
# To complete this concept it may take 40 hours to understand completely
# but for ML, DL, we will discuss now

# Pandas is a Data analysis library
# Extensively used by Data Scientists and Data Analysts..
# Pandas are works over and above the basic fundamental library called Numpy
# basically Numpy works inside of pandas becuase basic opreations which are available in pandas for numerical operations like we have seen in nUMPY, CAN BE A limitation in Pandas, So always pandas goes with Numpy also
# Numpy should be installed along with Pandas for better performance of numberical operations

# In python need to deal with dataFrames then pandas is very important lib.

# For EDA, Pandas becomes a keyrole 

# install Pandas
!pip install pandas

import numpy as np
import pandas as pd

# when we use statistical operations like central tendency parameters....Numpy is required, So numpy is ompotrtant in Pandas.
# even for hypothesis numpy is required

# Delimited values like comma (,) other formats pandas works well
# in case of csv format, delimted value is comma (,)

# The core abstraction layer in pandas is DataFrames
# DataFrames are created with a set of series elements

# In pandas basically we will encounter DataFrmes and series in genreal.
# should be known what is DataFrame and series?

# Creating a DataFrame
# 2 methods are there
# 1. Use a collection of objects (lists, dictionaries and Numpy)
# 2. Using file

# Creating a DF using list
empList = [[1,"Aravind",400000],[2,"Sharada",300000],[3,"Aaradhya",200000],[4,"Abhishek",150000],[5,"Aounnathya",100000]]

empList

print(empList)

# # creation of DF using pandas
empDataList = pd.DataFrame(empList)


empDataList

type(empDataList)

# Dataframe contains two types of indexes
# 1. column index and 2. Row index
# # assign column names to the list 
empDataList.columns = ["Eid","Ename","Esalary"]

empDataList

empDataList1 = pd.DataFrame(empList, columns= ["Eid","Ename","Esalary"])

empDataList1

# exact index should be mentioned

# DtaFrame is the official name of table in pandas. end of the story :-)

 # According to oops, DataFrame is the class which extracts the homogeneous data

empDataList.info

empDataList.info()

# In pandas , anything other than numeric is an object
# as per pythonic paradigm, string and objects both are same.

# row index is the range index

empList1 = [[1,"Aravind",400000],[2,"Sharada",300000],[3,"Aaradhya",200000],[4,"Abhishek",150000],[5,"Aounnathya",]]

empList1

empDataList2 = pd.DataFrame(empList1, columns= ["Eid","Ename","Esalary"])

empDataList2

# NaN is not a number

np.nan

empDataList1.info()

empList2 = [[1,"Aravind",400000],[2,"Sharada",300000],[3,"Aaradhya",200000],[4,"Abhishek",150000],[5,"Aounnathya",],[6, " ", 12000]]

empList2

empDataList3= pd.DataFrame(empList2, columns= ["Eid","Ename","Esalary"])

empDataList3

# the above data is having 1 nan value in esalary and 1 empty value in ename

# empty value in ename be corrected as NaN

empList3 = [[1,"Aravind",400000],[2,"Sharada",300000],[3,"Aaradhya",200000],[4,"Abhishek",150000],[5,"Aounnathya",],[6,np.nan, 12000]]

empList3

empDataList4= pd.DataFrame(empList3, columns= ["Eid","Ename","Esalary"])

empDataList4

# how to read NaN values, 
# 1 NaN value in ename and 1 NaN value in Esalary

empDataList4.info()

# How to read above data
# Range index is row index, which is having 6 entries from 0 to 5
# Column entries 3
# Observe the NUll values (NaN) in the columns
# Ename has 1 NaN
# Esalary has 1NaN
# did you understand HOW info is eye balling the missing value
# this is more like eye balling

# the above function is interpretation of data

# The above is based on the list

# Creating DF from file (Delimited file)

# Delimited refers to how column values are seperated
# which ASCII character is used to seperate column data
# CSV-- comma seperated values (,)
# TSV-- tab seperated values (\t)
# PSV--pipe seperated values (|)--not used in market, they say Pi seperated values
# SSV--Semi colan seperated values (;)
# The above format values are depends on the user/Data provider
# assume that data is Fetching/ extracting from the IOT Device ,
# In the IOT micro controller, they coded to seperate columns based on Pi
# should understand, in which format the data is available

# Pandas will help us to deal with delimited files easily

empData = pd.read_csv("employee.csv")

empData

# read_csv : even though _csv is there here, it will support all the formats of the files.
# CSV in pandas python typically demands having a header in file, 
# otherwise we need to mention saying header=None

empData = pd.read_csv("employee.csv", header=None)

empData

empData = pd.read_csv("employee.csv", header=None, names = ['Eid','Ename','Esalary'])

empData

empData1 = pd.read_csv("employee.csv", # path file
                       header=None,    # Defines whether header is there or not
                       names = ['Eid','Ename','Esalary'], # Assign Name of the columns
                      sep = ",") # Informing the seperator value ; defalut sep is ,

empData1

empData1.head()

empData1.tail()

empData1.info()

# Anything end with pandas is dataframe

# the moment we give names , then header will be override by names,
# if you give names, then header is not required

# if comma is there in the data, it will be treated as object , so it will not influence 

# any seperator issues will be chased by data provider itself..

# at the microcontroller level
# at API level 
# web API level the data will be controlled with formats.

# CR and LF are control characters or 
# bytecode that can be used to mark a line break in a text file.
# LF = Line Feed ( \n , 0x0A in hexadecimal, 10 in decimal)
#— moves the cursor down to the next line without returning to the beginning of the line

# use CRLF to denote end of line. 
# core. eol = lf When Git needs to change line endings to write
# a file in your working directory it will always use LF to denote end of line

 # CRLF and LF both consider as \n ?
 # Yes

# when you work with the data, data information must be there from data team,
# otherwise how will work with the data
# Data information is important and need to check before start working

# can we name the row header too? similar as columns; Yes



# to see intellisense 
# press tab
# eg. pd.arr..then press tab for intellisense

# in general, to extract the data from the APIs, 
# we need to use webscapping methods like beautiful soap
# whatever method we use to get the data
# we can give it to pandas with output of webscapping 
# with codes to identify the seperator value
# and then load it into pandas

# with pandas we can not do webscrapping

# here we html to read
pd.read_html # but very complex to use
# this method is not recommended, instead can be used Beatiful soap, 
# process the data, get the data, store at one place then use it with pandas to load the data

# then load for analysis..is a good idea.

# Softwares are there for Webscraping, which return data in xlsx or other formats.
# That can be passed to pandas
# beautiful soup and scrapy?
# scrapy is easy to use if you ask which one is best
# if you are good at selenium, then is a good approach for data extraction
# Uipath also can be used for webscapping
# RPI also can be used

# what is Uipath? 
# RPA is Uipath
# Uipath is a software wwhich will help us to perform robotic operations 0f automation.
# We can automate the process of repetative task that need to perform
# RPA is the concept and UPI Path is the software

# using remote access , we can load the load in to pandas
# https://gist.githubusercontent.com/curran/a08a1080b88344b0c8a7/raw/639388c2cbc2120a14dcf466e85730eb8be498bb/iris.csv
# if there is no security instruction, then accesing the data with remote access is easy
# observe the data is with comma seperated values
# this is iris dataset

irisDataFromRemote = pd.read_csv("https://gist.githubusercontent.com/curran/a08a1080b88344b0c8a7/raw/639388c2cbc2120a14dcf466e85730eb8be498bb/iris.csv")

irisDataFromRemote

irisDataFromRemote.info()

# memory usage is 6+KB 
# now this data is in the system memory
# Remember Pandas are etremeely memory intensive

# focus more on security part and what kind of format data in available
# when you are laoding the data from remote/internet

irisDataFromRemote.head()

irisDataFromRemote.tail()

irisDataFromRemote.head(10)

# like in SQL, top, here .head() function is used to retrive thr data 
# head and tail functions are used to retrieve first and last records
# which is same as in linux
