==> Pandas in python/2. Series.ipynb <==
# Series is important function in analsing the single column in the dataset

import numpy as np
import pandas as pd

# Why numpy is required when we do pandas for anaysis
# ans; we dont want to use experimental numpy
# pandas are used because which are made on top of numpy
# Important : numpy is helpful in pandas because of numerical operations
# If we won't install numpy, still we can get the detils from experimental numpy package
# But we can not rely on that; reliability is the concern
# thatswhy always recommend standard numpy package along with Pandas for better numerical operations
# remember numpy participate in Pandas when and only in performance of statistics

# Series
# is nothing but a column in a DataFrame

# empList = [[1,"Aravind",400000],[2,"Sharada",300000],[3,"Aaradhya",200000],[4,"Abhishek",150000],[5,"Aounnathya", 120000],[6,"Robo", 120000]]

empData = pd.read_csv("employee.csv", # path file
                       header=None,    # Defines whether header is there or not
                       names = ['Eid','Ename','Esalary'], # Assign Name of the columns
                      sep = ",") # Informing the seperator value ; defalut sep is ,

empData

empData.info()

# like in SQL 
# table.column_name
empData.Esalary

type(empData) # its a dataframe

type(empData.Esalary) # ITS ASERIES

# Dataframe is a collection of series
# Dataframe can store heteogeneous series
# But series is always homogeneous

# why we focusing more on series
# why we are learning series

# Whenever we do statical operations, we perform on series ideally
# practically it differs

# first observe measures of central tendency
# mean ,median, mode

# MEAN
empData.Esalary.mean() # this is same as in numpy..thats y here numpy is required

# 75th percentile
np.percentile(empData.Esalary,75)

# 50th percentile
np.percentile(empData.Esalary,50)

# we call , columns as attributes, 
# if any error in column names it throw an error as attribute error

# attribute error is column section error

empData.Esalary.median()

# mode can not be calculated in numpy 
# we need to use scipy

# But Pandas having scipy by default

empData.Esalary.mode()

# empData.Esalary.percentile(50) 

# Statical analysis, Numerical analysis and Data analysis
# can bedone using single package is pandas

# pandas is having scipy experimental package
# but still we can install scipy package also

import scipy as sp

# but scipy will be taught more when we use statistical testing

#then Measures of dispersion
# Variance and SD

empData.Esalary.var()

empData.Esalary.std()

# Standard deviation is very huge, can be seen in graph also
# Pandas also supports plots but later can be seen

# the above series of column numerical analysis can be seen in single shot
empData.describe()

# if you see the mean and #50%le values
# mean = 300000 and 50% is 300000 both are same so, its following normal distribution

# there is no outlier exist in the data

# Before analysing the data, you need to convert Dataframe into numpy array
# why we need to do this?
# Ans: many situations may arise, and few are here
# 

Situation-1:

# If you are writting a code to create a statistical model in python
# basically its inferential statisticalmodel
# will use a package called scikit learn (ML) and Tensorflow (DL)
# All the packages expects your data should be in the form of Numpy array
# after completion of EDA, before taking to the model.. it should be in Numpy array

Situation-2: 

# if you want to perform statistical testing
# (Hypothesis testing, AB Testing), chi-squre test...
# then formula expects your data in the form of Numpy
# this is the prospective of requirement of form of Numpy array data

# save header names to a variable as headers will be lost while converting 
# Dataframe to Numpy array
empDataHeaders = empData.columns
# now the data code is 

empData.values

# now the above data is in the form of array

# now store it in Numpy array name
empDataNumpyArray = empData.values

empDataNumpyArray

# One thing need to observe here is 
# we loose header data...

# How to convert Numpy array into dataframe
empData2 = pd.DataFrame(empDataNumpyArray, columns = empDataHeaders)

empData2

# what is the adv of .values over empData values
# Because .values (Numpy array) offered by Pandas for further analysis
# otherwise both are same

empData.describe()

# Is above data valid or invalid interms of the above values are 
# calculated only for numericals?
# Answer is subjective , because according to statistics , 
# which is deals with numericals only.
# according to data science , all types of data to be 
# considered for analysis.
# so Pandas package is developed very smartly to cover 
# all the types of data.
# But Pandas has given default option to only for numericals. 


empData.describe(include="all")

# being a Data Scientist , its your responsibility to interpret the above data
# Unique , top and freq details are given to only categorical datatypes.

# categorical data can be seen seperately like numerical data types
empData.describe(include = np.object)

# frequecy here of top details...here top is sharada and at only once available
# as there is no repeat names, then top will be according to the descending order of the alphabets 
# then which ever is on top will become top.

# Here nobody is arguing with statisticians, but we can go ahead with whatever
# perception we have about the data

# Pandas creators , created this for all mindset people.

# according to statistical analysis, only numerical analysis is possible.
# But Statistical anaysis is required for Data Scientist.
# So being a DataScientist, think about Data Analysis and unserstand the insight of the data; 
# So we found some technics to convert string/ object data types values to numerical, 
# then we can use statistical anaylysis.

# possiblity to represent the non-numeric  column as numeric. in other words just like ANSI number ?
# yes
# we use some methods to do this.
# NLP
# DUMMY VARIABLES

# (include="all") will give complete data, and in this many NaN values will not give sufficient insight.
# Better way is to use default option, such a way that numerical option can be dealt and
# (include= np.object) can be used treat string/object data type values.

# Normal distribution is used only for numerical data

# but for String/object, can check for uniqueness , and repeatativeness...
# can be seen with EDA analysis as and when proceed further.

# .describe() can be used only for DataFrame.

# empData.Esalary.discribe()

empData.Esalary.discribe()

pd._version_

empData.Esalary.discribe().to_frame()
