==> Pandas in python/3. Operations in series.ipynb <==
import numpy as np
import pandas as pd

empWithHeaderDF = pd.read_csv("employeewithheaders.csv")

empWithHeaderDF

# i want to calculate the yearly salary of each employee on basis of monthly.
# Add a new column named yearlySalary

empWithHeaderDF.Esalary*12 # element wise operation

type(empWithHeaderDF.Esalary*12)

empWithHeaderDF["EsalaryYearly"] = empWithHeaderDF.Esalary*12

empWithHeaderDF

# eSalaryYearly = empWithHeaderDF.Esalary*12 # element wise operation

# eSalaryYearly 

# type(eSalaryYearly)

# Pandas Dataframe is a mutable object

# Use-Case 2 : Bonus od the employee
#  New column bonus should be added
# Esalary less than or equal to 1500/- --> 10% increment
# if the salary between 1501 to 6000/----> 5% increment
# if salary between 6001 to 9000/- ------> 2.5% increment
# if salary greater than 9001/- ------------> No increment

# Function must be return updated salary from this month
# when you have multiple checkpoints/ operations to perform , creat function.

def newSalary(Esalary):
    newSalary=0
    
    if Esalary <= 1500:
        newSalary = Esalary * 1.1 # (salary + salary*10%)
    elif Esalary<=6000:
        newSalary = Esalary * 1.05
    elif Esalary<=9000:
        newSalary = Esalary * 1.025
    else:
        newSalary = Esalary
    return newSalary

newSalary(8900)

newSalary(5500)

# Pandas allows you to apply function on series

updatedSalary = empWithHeaderDF.Esalary.apply(newSalary)

updatedSalary

empWithHeaderDF["updatedMonthlySalary"] = updatedSalary

empWithHeaderDF

empWithHeaderDF["updatedSalaryYearly"] = empWithHeaderDF.updatedMonthlySalary*12


empWithHeaderDF

empWithHeaderDF["updatedSalaryYearly"] = empWithHeaderDF.Esalary.apply(newSalary)*12

empWithHeaderDF

# adding custom series in an existing dataframe
pd.Series(["HR",'Design','Operation',"PMG","Delivery","Testing"])

empWithHeaderDF["Dept"] = pd.Series(["HR",'Design','Operation',"PMG","Delivery","Testing"])

empWithHeaderDF

# Export the DF to a file
empWithHeaderDF.to_csv("Finaloutput of empWithHeaderDF1.csv")

# the out put will have defauly index values which 
# is independent column generates with the above code

# best practice is not to have independant index column
empWithHeaderDF.to_csv("Finaloutput of empWithHeaderDF.csv" , index=False)
