==> Pandas in python/4. Traversing DF.ipynb <==
import numpy as np
import pandas as pd

datasetExample = pd.read_csv("Finaloutput of empWithHeaderDF.csv")

datasetExample

type(datasetExample)

# Trasvering the DF means accesing the elements
# this is same as in numpy which is with fancy indeing, index loacation..etc
# In panndas, many number of technics are there to access the elements
# 1. .head(3)
# But restrict to only one method, and become master in that.

# access first 3 records
# this will happen with identification of index numbers

# iloc method = index location -> traverse elements on the basis of row index and column index 
# ( access elements)

# syntax
# variation-1 : iloc(row_index, column_index)
# variation-2 : iloc(row_range, column_range) # incase of slicing
# where row_range = start value: upperbound
# upperbound = end value+1

datasetExample1 = datasetExample.iloc[0:3,0:7]

datasetExample1

datasetExample2 = datasetExample.iloc[0:3,0:]

datasetExample2

datasetExample2.iloc[0][1]

# access first 2 rows with Ename,EsalaryYearly and Dept
datasetExample4 = datasetExample.iloc[0:2, [1,3,6]]

datasetExample4

# access first 2 rows with Esalary and EsalaryYearly
datasetExample3 = datasetExample.iloc[0:2, 2:4]
datasetExample3

# access first and last rows with Esalary and EsalaryYearly
datasetExample5 = datasetExample.iloc[[0,-1], 2:4]

datasetExample5

# there is another method is loc
