==> Pandas in python/5. Operations in DF.ipynb <==
# OPERATION-1: Relational, logical and comparison operation

import numpy as np
import pandas as pd

datasetExample = pd.read_csv("Finaloutput of empWithHeaderDF.csv")

datasetExample

# extract record of emp. whose salary is greater than 5000
datasetExample.Esalary>5000

datasetExample[datasetExample.Esalary>5000]

# extract record of Ename and Dept of emp. whose salary is greater than 5000
datasetExample.iloc[0:, [1,6]][datasetExample.Esalary>5000]

datasetExample[datasetExample.Esalary>5000].iloc[:, [1,6]]

# extract the data whose salary is > 8000 and belongs to Delivery team
datasetExample

datasetExample[(datasetExample.Esalary>8000) & (datasetExample.Dept=="Delivery")]

# tuple is compulsory when use logical contions for biwise operator

# list the names of the employees whose salary is >1900 and belongs to operation team
datasetExample

datasetExample.iloc[0:,1][(datasetExample.Esalary>1990) & (datasetExample.Dept=="Operation")]

datasetExample[(datasetExample.Esalary>1990) & (datasetExample.Dept=="Operation")].Ename

datasetExample[(datasetExample.Esalary>1990) & (datasetExample.Dept=="Operation")].iloc[0:,1]

datasetExample[(datasetExample.Esalary>1990) & (datasetExample.Dept=="Operation")].iloc[0:,1].to_frame()

# the functions are worked well with . operator..
# but if there is a space in the columns names like yearly salary instead of yearlysalary,
# there will be a problem 

# check with space once
datasetExample["area of working"] = [" Bangalore","Hyderabad","Bangalore","Hyderabad","Chennai","Kurnool"]

datasetExample

datasetExample.area of working

#.(dot) operator will not work if there is a space in the name of the column

# instead of . , we can use [""]
datasetExample["area of working"]

#everything is same except . operator

datasetExample["area of working"].to_frame()

datasetExample[datasetExample.area of working ==" Hyderabad"]

# .operator will throw an error when there is a space in the name of the column

datasetExample[datasetExample["area of working"]=="Hyderabad"]

datasetExample[datasetExample["area of working"]=="Hyderabad"]

# Operation-2: Dealing of duplicate rows

# duplicate records
# full records in the row should be same , then only it will become dupliactes and will be removed
datasetExample

datasetExample.drop_duplicates() # without commit in the real data frame

# here there are no dupliactes

# the above code is only to check for duplicates.. it will not commit to remove

datasetExample.drop_duplicates( inplace=True)  # inplace is an commit operation in database
# this will commit to change the records in DF

datasetExample # this is after removal of duplicates

# Whenever dupliactes get deleted, the the numbers of index will be affected..
# that is, that row row number as index number will be removed completely
# This index number affects the analysis..
# so to reset the index, we need to do thr reset index.
# next class we will see for reset index technic.

# when we say commiting the data to be removed...in python, obejects are stored in memory objects..
# so it will works on memory obects.
# right now its working on RAM..
# the moment close the jupyternotebook lab, it will be removed from the moemory objects.

datasetExample.to_csv("Output of dupliactes removed file.csv", index=False)
