==> Pandas in python/6. Join and merge operation.ipynb <==
# Operation 3: Performing merge operation in pandas (equal to join operation in database)

import numpy as np
import pandas as pd

datasetDF = pd.read_csv("Output of dupliactes removed file.csv")

datasetDF

# we creATE A NEW ddf
dataExampleDF1 = pd.DataFrame([ [4,"QA"],[1,"HR"],[2,"Ops"],[3,"Dev"] ], columns=["Eid","Dept."])

dataExampleDF1

# whatever way we have created, same way it is displayed, deliberately it has taken.

dataExampleDF2 = pd.DataFrame([ [1,"Sharada"],[2,"Aaradhya"],[3,"Abhishek"] ], columns=["Eid","Ename"])

dataExampleDF2

# from above 2 tables, common column is Eid.

# Task 1: I want to join above two datasets logically on the basis of Eid

# when you have common column in both tables, merge option will be used.

pd.merge(dataExampleDF1,dataExampleDF2)

# common items only will be displayed.

dataExampleDF3 = pd.DataFrame([ [4,"QA"],[1,"HR"],[2,"Ops"],[3,"Dev"] ], columns=["Eid","Dept."])
dataExampleDF4 = pd.DataFrame([ [1,"Sharada"],[2,"Aaradhya"],[3,"Abhishek"] ], columns=["Empid","Ename"])

pd.merge(dataExampleDF3,dataExampleDF4)

# becuase there is no common column, so megring will not happen.
# but, customised merging can be done
pd.merge(dataExampleDF3,dataExampleDF4, left_on="Eid" , right_on="Empid")

# only two tables can be merged. the resultant DF can be merged with another tablet.

# concat operation

dataEmpDF1 = pd.DataFrame([ [1,"Sharada"],[2,"Aaradhya"],[3,"Abhishek"] ], columns=["Eid","Ename"])
dataEmpDF2 = pd.DataFrame([ [4,"Aravind"],[5,"Aounnathya"],[6,"Robo"] ], columns=["Eid","Ename"])

resultEmpDF = pd.concat( [dataEmpDF1,dataEmpDF2] , axis=0) # axis=0 means row concat

resultEmpDF

# In many operations, indexing is the challenging as it requires reset index.

# many number of DFs/tables can be concatnated , but in merge operation only two DF can be merged.

### column wise concatnation

dataEmpDF3 = pd.DataFrame([ [4,"QA"],[1,"HR"],[2,"Ops"],[3,"Dev"] ], columns=["Eid","Dept."])
dataEmpDF4 = pd.DataFrame([ [1,"Sharada"],[2,"Aaradhya"],[3,"Abhishek"] ], columns=["Empid","Ename"])

resultEmpDF2 = pd.concat( [dataEmpDF3,dataEmpDF4] , axis=1) # axis=1 for columnwise concat
resultEmpDF2

# Empid is float here , its by default

# if there is a relationship between DF1 and DF2 , then cancat can not be used 
# here merge should be used.

# concat will not help you about relationship of Two DFs
# just it will paste the data at one place.
# if we are looking for Primary key and foreign key relationship concept ,such senerio can concat cn not be used

# column concat done based on row indexing

resultEmpDF2 = pd.concat( [dataEmpDF3,dataEmpDF4]) 
resultEmpDF2
