==> Pandas in python/7. Deletion operator.ipynb <==
import numpy as np
import pandas as pd

datasetDF = pd.read_csv("Output of dupliactes removed file.csv")

datasetDF

# delete Esalary and EsalaryYearly
datasetDF.drop(["Esalary","EsalaryYearly"], axis=1)

datasetDF

# the above drop/delete operation is not committed, means not permanemntly deleted
# this is only to check for drop details

# delete a column
# to drop completely even into DF then use inplace=True
datasetDF.drop(["Esalary","EsalaryYearly"], axis=1 , inplace=True)

datasetDF

# Delete a row
# in both the cases, drop can take only indexes.

datasetDF.drop([5], axis=0)

# here based on row index its dropped, ofcourse not committed
datasetDF

# How dynamically trace the indexes and to drop the same
datasetDF[datasetDF.Eid==5].index

datasetDF.drop(datasetDF[datasetDF.Eid==5].index , axis=0)

# truncate thre yable data
datasetDF.truncate(inplace=True)
